Replace implicit ignores with explicit ones
The pattern `result.is_ok();` only evades the use checker in earlier
version of Rust, however it does not clearly signal that the result
should be discarded. An assignment to the `_` pattern should be used
instead:

> let _ = result;

This is not a real binding, it drops the value of the right-hand-side
value immediately just like the by-value call to `Result::is_ok`.